refactor(task-manager): unify Prefect access behind one client adapter
Replace the per-component reader and retention client adapters, and the counter's direct private-client call, with a single set of capability-role protocols and one PrefectClientAdapter. Each consumer depends only on the roles it uses; the adapter is built once and injected into the reader, counter, and retention. Test fakes are narrowed to each consumer's roles.
